👌 IMPROVE: move gover-core to typescript
