+++ packages/govern-core/.env.example
@@ -1,3 +1,2 @@
-ETH_KEY=
 ETHERSCAN_KEY=
 REGISTRY_RINKEBY=0x874E22352098748504C8a8644b1492436CB73E52
