+++ packages/govern-core/.gitignore
@@ -10,3 +10,4 @@ artifacts
 coverage
 coverage.json
 .coverage_*
+typechain
