+++ packages/govern-core/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
+    "strict": true,
+    "esModuleInterop": true,
+    "outDir": "dist"
+  },
+  "include": ["./scripts", "./test"],
+  "files": [
+    "./buidler.config.ts",
+    "../../node_modules/@nomiclabs/buidler-ethers/src/type-extensions.d.ts",
+    "../../node_modules/@nomiclabs/buidler-waffle/src/type-extensions.d.ts",
+    "../../node_modules/@nomiclabs/buidler-etherscan/src/type-extensions.d.ts",
+    "../../node_modules/buidler-typechain/src/type-extensions.d.ts",
+    "../../node_modules/buidler-local-networks-config-plugin/src/type-extensions.d.ts"
+  ]
+}
